Will create actions and reduser
