+++ src/const.ts
@@ -1,3 +1,5 @@
+import { CityLocations } from './types/types.ts';
+
 const REVIEW_COUNT = 5;
 const AVATAR_URL = 'https://i.pravatar.cc/128';
 const URL_MARKER_DEFAULT = 'markup/img/pin.svg';
@@ -22,6 +24,16 @@ enum AuthorizationStatus {
 }
 
 const CITIES = ['Paris', 'Cologne', 'Brussels', 'Amsterdam', 'Hamburg', 'Dusseldorf'];
+
+const CITY_LOCATIONS: CityLocations = {
+  Paris: { latitude: 48.85661, longitude: 2.351499, zoom: 13 },
+  Cologne: { latitude: 50.938361, longitude: 6.959974, zoom: 13 },
+  Brussels: { latitude: 50.846557, longitude: 4.351697, zoom: 13 },
+  Amsterdam: { latitude: 52.37454, longitude: 4.897976, zoom: 13 },
+  Hamburg: { latitude: 53.550341, longitude: 10.000654, zoom: 13 },
+  Dusseldorf: { latitude: 51.225402, longitude: 6.776314, zoom: 13 },
+};
+
 const TYPES = ['Room', 'Apartment', 'House', 'Hotel'];
 
 const PLACES_OPTIONS = ['Popular', 'Price: low to high', 'Price: high to low', 'Top rated first'];
@@ -43,4 +55,5 @@ export {
   TILE_LAYER_ATTRIBUTION,
   MIN_REVIEW_LENGTH,
   MAX_REVIEW_LENGTH,
+  CITY_LOCATIONS,
 };
